editing XOR and referencing the processing core library
The XOR example was written in processing with a Graphical Interface
using a lot of the processing functions. And since it would have been
quite some work to write a fully new version using only the standard
JRE, I simply referenced core.jar of the processing library to the
project.